requirements.txt
imap-tools
python-dotenv
python-telegram-bot
PyYAML
beautifulsoup4